Prehrani testovacich verzi balicku

+++ elza/elza-core/src/main/java/cz/tacr/elza/packageimport/PackageService.java
@@ -577,7 +577,7 @@ public class PackageService {
             }
         }
 
-        if (oldPackageDir != null) {
+        if (oldPackageDir != null && !oldPackageDir.equals(packageDir)) {
             FileSystemUtils.deleteRecursively(oldPackageDir);
         }
